# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tugas REST API Express.js + Supabase & JWT",
    "version": "1.0.0",
    "description": "Dokumentasi REST API interaktif menggunakan Swagger UI. API ini dilengkapi autentikasi JWT dan database Supabase."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server Pengembangan Lokal"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Registrasi Pengguna Baru",
        "description": "Mendaftarkan akun pengguna baru ke sistem dengan melakukan hashing password terlebih dahulu menggunakan bcryptjs.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "budi@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "rahasiaBudi123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Registrasi berhasil"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Input tidak valid atau email sudah terdaftar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login Pengguna",
        "description": "Memverifikasi email dan password untuk mendapatkan token JWT agar bisa mengakses rute yang dilindungi.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "budi@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "rahasiaBudi123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login berhasil"
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Email atau password salah",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles": {
      "get": {
        "summary": "Mendapatkan Semua Artikel",
        "description": "Menampilkan daftar seluruh artikel yang ada di database. Memerlukan autentikasi Bearer Token JWT.",
        "tags": ["Articles"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil mendapatkan semua artikel",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Berhasil mendapatkan semua artikel"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Article"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token tidak disediakan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Token tidak valid atau kedaluwarsa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Membuat Artikel Baru",
        "description": "Menambahkan artikel baru ke database. Field `author_id` akan diisi secara otomatis dari ID user yang masuk (didapatkan dari token JWT).",
        "tags": ["Articles"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Tutorial Pemrograman Express.js"
                  },
                  "content": {
                    "type": "string",
                    "example": "Express.js adalah web framework minimalis untuk Node.js..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artikel berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Artikel berhasil dibuat"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Article"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Title atau content kosong",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "summary": "Mendapatkan Artikel Berdasarkan ID",
        "description": "Menampilkan artikel secara mendalam berdasarkan ID uniknya. Memerlukan autentikasi Bearer Token JWT.",
        "tags": ["Articles"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID unik Artikel (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil mendapatkan artikel",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Berhasil mendapatkan artikel"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Article"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artikel tidak ditemukan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "e4277717-38de-4fae-9d21-4fa35c4bf9bd"
          },
          "name": {
            "type": "string",
            "example": "Budi Santoso"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "budi@example.com"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-25T12:00:00.000Z"
          }
        }
      },
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "f5b6a7c8-d9e0-1f2a-3b4c-5d6e7f8g9h0i"
          },
          "title": {
            "type": "string",
            "example": "Tutorial Pemrograman Express.js"
          },
          "content": {
            "type": "string",
            "example": "Express.js adalah web framework minimalis untuk Node.js..."
          },
          "author_id": {
            "type": "string",
            "format": "uuid",
            "example": "e4277717-38de-4fae-9d21-4fa35c4bf9bd"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-25T12:10:00.000Z"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Deskripsi pesan kesalahan dari server"
          }
        }
      }
    }
  }
}